Ported to KDE 3.0, added beginnings for Drag 'n' Drop classes
svn path=/trunk/kdenlive/; revision=20

+++ kdenlive/kmmruler.cpp
@@ -122,7 +122,7 @@ void KMMRuler::setFrameSize(double size){
 	
 }
 
-int KMMRuler::getTickCount(int size)
+int KMMRuler::getTickCount(double size)
 {
 	int tick=1;
